Remove leading slash
This happens a lot when I'm developing. It works in development but fails in production. This is because I'm always serving my development assets from the project's root location. When I move to production, my assets are not served from the project's root so it causes a 404 error when loading the libraries.